model.py
import os

import gspread
from dotenv import load_dotenv
from google.oauth2.service_account import Credentials


def authorize_gspread():
    load_dotenv()
    SCOPES = ['https://www.googleapis.com/auth/spreadsheets']
    SERVICE_ACCOUNT_FILE = os.getenv('SERVICE_ACCOUNT_FILE')
    creds = Credentials.from_service_account_file(
        SERVICE_ACCOUNT_FILE, scopes=SCOPES)
    gc = gspread.authorize(creds)
    return gc
