pyproject.toml
[tool.poetry]
name = "newsletter"
version = "0.1.0"
description = ""
authors = ["TMREasonlee <easonlee@tmrmds.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
gspread = "^6.2.1"
google-auth = "^2.40.3"
pandas = "^2.3.0"
python-dotenv = "^1.1.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
